[feat] Enable PartFinal tests on Windows
Add Linux-compatible TBOR OOB metadata to the Windows DDI and consolidate PartFinal coverage into backend-neutral tests with updated coverage documentation.

Copilot-Session: 7ccbdc1e-2946-4f0c-b020-b3f1ade57f77
